fix: static site generation with regions
