+++ frontend/utils/constants.ts
@@ -1,0 +1,6 @@
+export const regions = [
+  { id: "kamppi", label: "Helsinki - Kamppi" },
+  { id: "tampere", label: "Tampere" },
+];
+
+export const defaultRegion = "kamppi";
